# my-prop/src/api/generated/domains/domains.msw.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */
import { faker } from "@faker-js/faker"

import { HttpResponse, http } from "msw"
import type { RequestHandlerOptions } from "msw"

import type { Domain } from "../model"

export const getListDomainsResponseMock = (): Domain[] =>
  Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => ({
    id: faker.string.uuid(),
    domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
    websiteId: faker.string.uuid(),
    websiteName: faker.string.alpha({ length: { min: 10, max: 20 } }),
    status: faker.helpers.arrayElement(["active", "pending"] as const),
    dnsRecords: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => ({
      type: faker.helpers.arrayElement(["A", "CNAME"] as const),
      name: faker.string.alpha({ length: { min: 10, max: 20 } }),
      value: faker.string.alpha({ length: { min: 10, max: 20 } }),
    })),
    createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  }))

export const getCreateDomainResponseMock = (
  overrideResponse: Partial<Extract<Domain, object>> = {}
): Domain => ({
  id: faker.string.uuid(),
  domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
  websiteId: faker.string.uuid(),
  websiteName: faker.string.alpha({ length: { min: 10, max: 20 } }),
  status: faker.helpers.arrayElement(["active", "pending"] as const),
  dnsRecords: Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => ({
    type: faker.helpers.arrayElement(["A", "CNAME"] as const),
    name: faker.string.alpha({ length: { min: 10, max: 20 } }),
    value: faker.string.alpha({ length: { min: 10, max: 20 } }),
  })),
  createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  ...overrideResponse,
})

export const getGetDomainResponseMock = (
  overrideResponse: Partial<Extract<Domain, object>> = {}
): Domain => ({
  id: faker.string.uuid(),
  domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
  websiteId: faker.string.uuid(),
  websiteName: faker.string.alpha({ length: { min: 10, max: 20 } }),
  status: faker.helpers.arrayElement(["active", "pending"] as const),
  dnsRecords: Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => ({
    type: faker.helpers.arrayElement(["A", "CNAME"] as const),
    name: faker.string.alpha({ length: { min: 10, max: 20 } }),
    value: faker.string.alpha({ length: { min: 10, max: 20 } }),
  })),
  createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  ...overrideResponse,
})

export const getVerifyDomainResponseMock = (
  overrideResponse: Partial<Extract<Domain, object>> = {}
): Domain => ({
  id: faker.string.uuid(),
  domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
  websiteId: faker.string.uuid(),
  websiteName: faker.string.alpha({ length: { min: 10, max: 20 } }),
  status: faker.helpers.arrayElement(["active", "pending"] as const),
  dnsRecords: Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => ({
    type: faker.helpers.arrayElement(["A", "CNAME"] as const),
    name: faker.string.alpha({ length: { min: 10, max: 20 } }),
    value: faker.string.alpha({ length: { min: 10, max: 20 } }),
  })),
  createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  ...overrideResponse,
})

export const getListDomainsMockHandler = (
  overrideResponse?:
    | Domain[]
    | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0]
      ) => Promise<Domain[]> | Domain[]),
  options?: RequestHandlerOptions
) => {
  return http.get(
    "/api/domains",
    async (info: Parameters<Parameters<typeof http.get>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getListDomainsResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}

export const getCreateDomainMockHandler = (
  overrideResponse?:
    | Domain
    | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0]
      ) => Promise<Domain> | Domain),
  options?: RequestHandlerOptions
) => {
  return http.post(
    "/api/domains",
    async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getCreateDomainResponseMock(),
        { status: 201 }
      )
    },
    options
  )
}

export const getGetDomainMockHandler = (
  overrideResponse?:
    | Domain
    | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0]
      ) => Promise<Domain> | Domain),
  options?: RequestHandlerOptions
) => {
  return http.get(
    "/api/domains/:domainId",
    async (info: Parameters<Parameters<typeof http.get>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getGetDomainResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}

export const getDeleteDomainMockHandler = (
  overrideResponse?:
    | void
    | ((
        info: Parameters<Parameters<typeof http.delete>[1]>[0]
      ) => Promise<void> | void),
  options?: RequestHandlerOptions
) => {
  return http.delete(
    "/api/domains/:domainId",
    async (info: Parameters<Parameters<typeof http.delete>[1]>[0]) => {
      if (typeof overrideResponse === "function") {
        await overrideResponse(info)
      }

      return new HttpResponse(null, { status: 204 })
    },
    options
  )
}

export const getVerifyDomainMockHandler = (
  overrideResponse?:
    | Domain
    | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0]
      ) => Promise<Domain> | Domain),
  options?: RequestHandlerOptions
) => {
  return http.post(
    "/api/domains/:domainId/verify",
    async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getVerifyDomainResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}
export const getDomainsMock = () => [
  getListDomainsMockHandler(),
  getCreateDomainMockHandler(),
  getGetDomainMockHandler(),
  getDeleteDomainMockHandler(),
  getVerifyDomainMockHandler(),
]